setupsdk: Warn the user they must source (.) the script, and fix dash
Rework the script to be more like oe-init-build-env.  We warn the user
if we can detect they ran this vs source it.  Additionally, dash does
not support 'source', so we much use the posix '.' and tell
oe-init-build-env how to locate itself.

Additionally, in case of a failure do not 'exit', as this can terminate
the shell if the error occurs when sourced.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>